Inject APIClient when initializing viewcontroller
